workshop done

+++ server/index.js
@@ -20,10 +20,32 @@ const path = require('path');
 app.get('/', (req, res)=> res.sendFile(path.join(__dirname, '../client/dist/index.html')));
 app.use('/assets', express.static(path.join(__dirname, '../client/dist/assets'))); 
 
+async function isLoggedIn(req, res, next){
+  try {
+    const user = await findUserWithToken(req.headers.authorization);
+    if(user){
+      req.user = user;
+      next();
+    }
+   else {
+    const error = Error('not authorized');
+    error.status = 401;
+    next(error)
+  }
+}
+catch(ex){
+  next(ex);
+}
+}
 
 app.post('/api/auth/login', async(req, res, next)=> {
   try {
-    res.send(await authenticate(req.body));
+    const { validUser, token } = await authenticate(req.body);
+   if (validUser) {
+    res.send({ token });
+   } else {
+    res.status(401).send({ error: 'not authorized'});
+   }
   }
   catch(ex){
     next(ex);
@@ -48,8 +70,13 @@ app.get('/api/users', async(req, res, next)=> {
   }
 });
 
-app.get('/api/users/:id/favorites', async(req, res, next)=> {
+app.get('/api/users/:id/favorites', isLoggedIn, async(req, res, next)=> {
   try {
+    if (req.user.id !== req.params.userId) {
+      const error = Error('not authorized');
+      error.status = 401;
+      throw error;
+    }  
     res.send(await fetchFavorites(req.params.id));
   }
   catch(ex){
@@ -57,8 +84,13 @@ app.get('/api/users/:id/favorites', async(req, res, next)=> {
   }
 });
 
-app.post('/api/users/:id/favorites', async(req, res, next)=> {
+app.post('/api/users/:id/favorites', isLoggedIn, async(req, res, next)=> {
   try {
+    if (req.user.id !== req.params.userId) {
+      const error = Error('not authorized');
+      error.status = 401;
+      throw error;
+    }  
     res.status(201).send(await createFavorite({ user_id: req.params.id, product_id: req.body.product_id}));
   }
   catch(ex){
@@ -66,9 +98,14 @@ app.post('/api/users/:id/favorites', async(req, res, next)=> {
   }
 });
 
-app.delete('/api/users/:user_id/favorites/:id', async(req, res, next)=> {
+app.delete('/api/users/:user_id/favorites/:id', isLoggedIn, async(req, res, next)=> {
   try {
-    await destroyFavorite({user_id: req.params.user_id, id: req.params.id });
+    if (req.user.id !== req.params.userId) {
+      const error = Error('not authorized');
+      error.status = 401;
+      throw error;
+    }   
+     await destroyFavorite({user_id: req.params.user_id, id: req.params.id });
     res.sendStatus(204);
   }
   catch(ex){
